feat: check pid alive before really killing it

+++ apps/electron-client/src/runtime.ts
@@ -3,7 +3,6 @@ import path from "path";
 
 import fs from "fs-extra";
 
-
 import { nrm } from "./nrm";
 import { logDir } from "./utils/commonPaths";
 
@@ -12,6 +11,15 @@ let runtimeProcess: ChildProcess | undefined;
 const logPath = path.join(logDir, "runtime.log");
 const errPath = path.join(logDir, "runtime.err.log");
 const pidPath = path.join(logDir, "runtime.pid");
+
+function pidIsRunning(pid: number): boolean {
+  try {
+    process.kill(pid, 0);
+    return true;
+  } catch (e) {
+    return false;
+  }
+}
 
 const getRuntimePath = (): string => {
   if (process.env.NODE_ENV === "development") {
@@ -24,10 +32,14 @@ const getRuntimePath = (): string => {
 const checkAndCloseExistingRuntime = (): void => {
   if (fs.existsSync(pidPath)) {
     const pid = fs.readFileSync(pidPath, "utf-8");
-    try {
-      process.kill(parseInt(pid));
-    } catch (e) {
-      console.log("Failed to kill runtime process", e);
+    if (pidIsRunning(parseInt(pid))) {
+      console.log("Killing existing runtime process", pid);
+
+      try {
+        process.kill(parseInt(pid));
+      } catch (e) {
+        console.log("Failed to kill runtime process", e);
+      }
     }
     fs.unlinkSync(pidPath);
   }
@@ -48,7 +60,7 @@ export const startRuntime = async () => {
     env: {
       ...process.env,
       PATH: `${binPath}:${process.env.PATH}`,
-    }
+    },
   });
 
   fs.writeFileSync(pidPath, runtimeProcess.pid.toString());
@@ -74,5 +86,4 @@ export const stopRuntime = (): void => {
 export const restartRuntime = () => {
   stopRuntime();
   startRuntime();
-}
-
+};
